% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility-matrix.R
\name{createLabelMatrix}
\alias{createLabelMatrix}
\title{Create the dataframe with event names and the related labels}
\usage{
createLabelMatrix(data, output_dir = NULL)
}
\arguments{
\item{data}{data structure as the files in "UCR Time Series Classification Archive"}

\item{output_dir}{the file to save the label matrix data}
}
\value{
the dataframe, including event names and labels
}
\description{
Create the dataframe with event names and the related labels
}
\examples{
\donttest{
original_data <- data.frame("1"=c(1, 2, 1), "2"=c(5,6,7), 
   "3"=c(4, 5, 8), "4"=c(3, 1, 9))
label_matrix <- createLabelMatrix(data = original_data)
}
}
