% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility-matrix.R
\name{createStandardMatrix}
\alias{createStandardMatrix}
\title{Create the dataframe, only including the event data}
\usage{
createStandardMatrix(data, output_dir = NULL)
}
\arguments{
\item{data}{data structure as the files in "UCR Time Series Classification Archive"}

\item{output_dir}{the file to save the standard matrix data}
}
\value{
the dataframe of event data
}
\description{
Create the dataframe, only including the event data
}
\examples{
\donttest{
original_data <- data.frame("1"=c(1, 2, 1), "2"=c(5,6,7), 
   "3"=c(4, 5, 8), "4"=c(3, 1, 9))
standard_matrix <- createStandardMatrix(data = original_data)
}
}
